extract general text generation utils to a separate package
