+++ packages/textgen-utils/.gitignore
@@ -1,0 +1,2 @@
+/dist
+/lionweb-json-*.tgz
