+++ packages/textgen-utils/.npmignore
@@ -1,0 +1,3 @@
+/src
+/tsconfig.json
+/lionweb-json-util-*.tgz
