+++ packages/textgen-utils/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## 0.7.0
+
+Initial creation and publication of this package, as an extraction from `@lionweb/ts-utils`.
+
