+++ packages/textgen-utils/README.md
@@ -1,0 +1,26 @@
+# The `textgen-utils` package
+
+[![license](https://img.shields.io/badge/License-Apache%202.0-green.svg?style=flat)
+](./LICENSE)
+[![CI](https://github.com/LionWeb-io/lionweb-typescript/actions/workflows/test.yaml/badge.svg)
+](https://github.com/LionWeb-io/lionweb-typescript/actions/workflows/test.yaml)
+[![npm](https://img.shields.io/npm/v/%40lionweb%2Fjson?label=%40lionweb%2Ftextgen-utils)
+](https://www.npmjs.com/package/@lionweb/textgen-utils)
+
+An NPM package that can be added to a Node.js/NPM codebase as follows:
+
+```shell
+$ npm add @lionweb/textgen-utils
+```
+
+It contains general utilities for doing text – i.e.: code – generation, typically based on the `littoral-templates` package.
+
+
+## Development
+
+Build it from source as follows:
+
+```shell
+npm run build
+```
+
