+++ packages/textgen-utils/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@lionweb/textgen-utils",
+  "version": "0.7.0-beta.1",
+  "description": "Utilities for text generation",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "typings": "dist/index.d.ts",
+  "type": "module",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/LionWeb-io/lionweb-typescript.git"
+  },
+  "bugs": {
+    "url": "https://github.com/LionWeb-io/lionweb-typescript/issues"
+  },
+  "scripts": {
+    "clean": "rm -rf dist/ node_modules/ lionweb-textgen-utils-*.tgz",
+    "build": "tsc",
+    "lint": "eslint src",
+    "prep:pre-release": "npm run clean && npm install && npm run build",
+    "prerelease-alpha": "npm run prep:pre-release",
+    "release-alpha": "npm publish --tag alpha",
+    "prerelease-beta": "npm run prep:pre-release",
+    "release-beta": "npm publish --tag beta",
+    "prerelease": "npm run prep:pre-release",
+    "release": "npm publish"
+  },
+  "dependencies": {
+    "littoral-templates": "0.3.0"
+  }
+}
